+++ jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  clearMocks: true,
-  setupFiles: ["<rootDir>/test_setup.js"],
-};