+++ src/download.ts
@@ -1,8 +1,3 @@
-/* eslint-disable no-unused-vars */
-
-import axios from "axios";
-
-import { SkynetClient } from "./client.js";
 import {
   addUrlQuery,
   defaultOptions,
@@ -11,7 +6,7 @@ import {
   trimUriPrefix,
   uriHandshakePrefix,
   uriHandshakeResolverPrefix,
-} from "./utils.js";
+} from "./utils";
 
 const defaultDownloadOptions = {
   ...defaultOptions("/"),
@@ -32,7 +27,7 @@ const defaultResolveHnsOptions = {
  * @param {Object} [customOptions.query] - A query object to convert to a query parameter string and append to the skylink.
  * @returns {string} - The full URL that was used.
  */
-SkynetClient.prototype.downloadFile = function (skylink, customOptions = {}) {
+export function downloadFile(skylink: string, customOptions = {}): string {
   const opts = { ...defaultDownloadOptions, ...this.customOptions, ...customOptions, download: true };
   const url = this.getSkylinkUrl(skylink, opts);
 
@@ -40,7 +35,7 @@ SkynetClient.prototype.downloadFile = function (skylink, customOptions = {}) {
   window.location = url;
 
   return url;
-};
+}
 
 /**
  * Initiates a download of the content of the skylink at the Handshake domain.
@@ -50,7 +45,7 @@ SkynetClient.prototype.downloadFile = function (skylink, customOptions = {}) {
  * @param {Object} [customOptions.query] - A query object to convert to a query parameter string and append to the URL.
  * @returns {string} - The full URL that was used.
  */
-SkynetClient.prototype.downloadFileHns = function (domain, customOptions = {}) {
+export async function downloadFileHns(domain: string, customOptions = {}): Promise<string> {
   const opts = { ...defaultDownloadHnsOptions, ...this.customOptions, ...customOptions, download: true };
   const url = this.getHnsUrl(domain, opts);
 
@@ -58,15 +53,16 @@ SkynetClient.prototype.downloadFileHns = function (domain, customOptions = {}) {
   window.location = url;
 
   return url;
-};
+}
 
-SkynetClient.prototype.getSkylinkUrl = function (skylink, customOptions = {}) {
+export function getSkylinkUrl(skylink: string, customOptions = {}): string {
   const opts = { ...defaultDownloadOptions, ...this.customOptions, ...customOptions };
   const query = opts.query ?? {};
   if (opts.download) {
     query.attachment = true;
   }
 
+  // URL-encode the path.
   let path = "";
   if (opts.path) {
     if (typeof opts.path != "string") {
@@ -75,15 +71,15 @@ SkynetClient.prototype.getSkylinkUrl = function (skylink, customOptions = {}) {
     // Encode each element of the path separately and join them.
     path = opts.path
       .split("/")
-      .map((element) => encodeURIComponent(element))
+      .map((element: string) => encodeURIComponent(element))
       .join("/");
   }
 
   const url = makeUrl(this.portalUrl, opts.endpointPath, parseSkylink(skylink), path);
   return addUrlQuery(url, query);
-};
+}
 
-SkynetClient.prototype.getHnsUrl = function (domain, customOptions = {}) {
+export function getHnsUrl(domain: string, customOptions = {}): string {
   const opts = { ...defaultDownloadHnsOptions, ...this.customOptions, ...customOptions };
   const query = opts.query ?? {};
   if (opts.download) {
@@ -92,23 +88,19 @@ SkynetClient.prototype.getHnsUrl = function (domain, customOptions = {}) {
 
   const url = makeUrl(this.portalUrl, opts.endpointPath, trimUriPrefix(domain, uriHandshakePrefix));
   return addUrlQuery(url, query);
-};
+}
 
-SkynetClient.prototype.getHnsresUrl = function (domain, customOptions = {}) {
+export function getHnsresUrl(domain: string, customOptions = {}): string {
   const opts = { ...defaultResolveHnsOptions, ...this.customOptions, ...customOptions };
-  const query = opts.query ?? {};
-  if (opts.download) {
-    query.attachment = true;
-  }
 
   return makeUrl(this.portalUrl, opts.endpointPath, trimUriPrefix(domain, uriHandshakeResolverPrefix));
-};
+}
 
-SkynetClient.prototype.getMetadata = async function (skylink, customOptions = {}) {
+export async function getMetadata(skylink: string, customOptions = {}) {
   const opts = { ...defaultDownloadOptions, ...this.customOptions, ...customOptions };
 
   throw new Error("Unimplemented");
-};
+}
 
 /**
  * Opens the content of the skylink within the browser.
@@ -116,14 +108,14 @@ SkynetClient.prototype.getMetadata = async function (skylink, customOptions = {}
  * @param {Object} [customOptions={}] - Additional settings that can optionally be set. See `downloadFile` for the full list.
  * @returns {string} - The full URL that was used.
  */
-SkynetClient.prototype.openFile = function (skylink, customOptions = {}) {
+export function openFile(skylink: string, customOptions = {}): string {
   const opts = { ...defaultDownloadOptions, ...this.customOptions, ...customOptions };
   const url = this.getSkylinkUrl(skylink, opts);
 
   window.open(url, "_blank");
 
   return url;
-};
+}
 
 /**
  * Opens the content of the skylink from the given Handshake domain within the browser.
@@ -131,7 +123,7 @@ SkynetClient.prototype.openFile = function (skylink, customOptions = {}) {
  * @param {Object} [customOptions={}] - Additional settings that can optionally be set. See `downloadFileHns` for the full list.
  * @returns {string} - The full URL that was used.
  */
-SkynetClient.prototype.openFileHns = function (domain, customOptions = {}) {
+export async function openFileHns(domain: string, customOptions = {}): Promise<string> {
   const opts = { ...defaultDownloadHnsOptions, ...this.customOptions, ...customOptions };
   const url = this.getHnsUrl(domain, opts);
 
@@ -139,7 +131,7 @@ SkynetClient.prototype.openFileHns = function (domain, customOptions = {}) {
   window.open(url, "_blank");
 
   return url;
-};
+}
 
 /**
  * @param {string} domain - Handshake resolver domain.
@@ -147,7 +139,7 @@ SkynetClient.prototype.openFileHns = function (domain, customOptions = {}) {
  * @param {string} [customOptions.endpointPath="/hnsres"] - The relative URL path of the portal endpoint to contact.
  * @param {Object} [customOptions.query] - A query object to convert to a query parameter string and append to the URL.
  */
-SkynetClient.prototype.resolveHns = async function (domain, customOptions = {}) {
+export async function resolveHns(domain: string, customOptions = {}): Promise<any> {
   const opts = { ...defaultResolveHnsOptions, ...this.customOptions, ...customOptions };
   const url = this.getHnsresUrl(domain, opts);
 
@@ -159,4 +151,4 @@ SkynetClient.prototype.resolveHns = async function (domain, customOptions = {}) 
   });
 
   return response.data;
-};
+}